backend/app/optpro_rag_kg/requirements.txt
flask==3.0.3
pydantic==2.8.2
python-dotenv==1.0.1
neo4j==5.23.1
pinecone-client==5.0.1
sentence-transformers==3.0.1
numpy==1.26.4
pdfminer.six==20240706
rapidfuzz==3.9.7
